+++ sklego/meta/__init__.py
@@ -1,0 +1,17 @@
+__all__ = [
+    'ConfusionBalancer',
+    'DecayEstimator',
+    'EstimatorTransformer',
+    'GroupedEstimator',
+    'OutlierRemover',
+    'SubjectiveClassifier',
+    'Thresholder',
+]
+
+from .confusion_balancer import ConfusionBalancer
+from .decay_estimator import DecayEstimator
+from .estimator_transformer import EstimatorTransformer
+from .grouped_estimator import GroupedEstimator
+from .outlier_remover import OutlierRemover
+from .subjective_classifier import SubjectiveClassifier
+from .thresholder import Thresholder
